requirements.txt
matplotlib==3.5.2
mysql-connector-python==8.0.29
nltk==3.7
numpy==1.23.1
pandas==1.4.3
PyMySQL==1.0.2
python-dotenv==0.20.0
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
sklearn==0.0
SQLAlchemy==1.4.39
streamlit==1.10.0
textblob==0.17.1
tweepy==4.10.0
wordcloud==1.8.2.2
lxml==4.9.1

